requirements.txt
astroid==2.12.13
black==22.12.0
certifi==2022.9.24
click==8.1.3
dill==0.3.6
distlib==0.3.6
filelock==3.8.0
isort==5.10.1
lazy-object-proxy==1.8.0
mccabe==0.7.0
mypy-extensions==0.4.3
mysql-connector-python==8.0.31
numpy==1.24.0
openexchangerate==1.5.5
pandas==1.5.2
pathspec==0.10.2
platformdirs==2.5.4
protobuf==3.20.1
pylint==2.15.8
PyMySQL==1.0.2
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
python-dateutil==2.8.2
pytz==2022.7
six==1.16.0
tomlkit==0.11.6
virtualenv==20.17.0
wrapt==1.14.1
